add deciaml_places to model Auction

+++ auction/models.py
@@ -24,8 +24,8 @@ class Product(models.Model):
 class Auction(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     number_of_bids = models.IntegerField()
-    base_price = models.FloatField(null=True)
-    final_price = models.FloatField(null=True)
+    base_price = models.DecimalField(null=True, max_digits=6, decimal_places=2)
+    final_price = models.DecimalField(null=True, max_digits=6, decimal_places=2)
     expired = models.BooleanField(default=False)
     time_starting = models.DateTimeField()
     time_ending = models.DateTimeField(null=True, blank=True)
